refactor: Reorganize project
