+++ app/src/main/java/me/worric/kotlinplayground/extensions/DelegatesExtensions.kt
@@ -1,4 +1,4 @@
-package me.worric.kotlinplayground
+package me.worric.kotlinplayground.extensions
 
 import kotlin.reflect.KProperty
